moved macros out of public, updated ENXIO warning
