+++ include/pros/ext_adi.h
@@ -47,7 +47,7 @@ namespace c {
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports.
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -64,7 +64,7 @@ adi_port_config_e_t ext_adi_port_get_config(uint8_t smart_port, uint8_t adi_port
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports.
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -81,7 +81,7 @@ int32_t ext_adi_port_get_value(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports.
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -103,7 +103,7 @@ int32_t ext_adi_port_set_config(uint8_t smart_port, uint8_t adi_port, adi_port_c
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO  - The given value is not within the range of ADI Ports.
+ * ENXIO  - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -134,7 +134,7 @@ int32_t ext_adi_port_set_value(uint8_t smart_port, uint8_t adi_port, int32_t val
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports.
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -153,7 +153,7 @@ int32_t ext_adi_analog_calibrate(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports.
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range.
  * EADDRINUSE - The port is not configured as an analog input
  *
  * \param smart_port
@@ -177,7 +177,7 @@ int32_t ext_adi_analog_read(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an analog input
  *
  * \param smart_port
@@ -206,7 +206,7 @@ int32_t ext_adi_analog_read_calibrated(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an analog input
  *
  * \param smart_port
@@ -230,7 +230,7 @@ int32_t ext_adi_analog_read_calibrated_HR(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a digital input
  *
  * \param smart_port
@@ -255,7 +255,7 @@ int32_t ext_adi_digital_read(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a digital input
  *
  * \param smart_port
@@ -275,7 +275,7 @@ int32_t ext_adi_digital_get_new_press(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a digital output
  *
  * \param smart_port
@@ -296,7 +296,7 @@ int32_t ext_adi_digital_write(uint8_t smart_port, uint8_t adi_port, bool value);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  *
  * \param smart_port
  *        The smart port number that the ADI Expander is in
@@ -315,7 +315,7 @@ int32_t ext_adi_pin_mode(uint8_t smart_port, uint8_t adi_port, uint8_t mode);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an motor
  *
  * \param smart_port
@@ -336,7 +336,7 @@ int32_t ext_adi_motor_set(uint8_t smart_port, uint8_t adi_port, int8_t speed);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an motor
  *
  * \param smart_port
@@ -353,7 +353,7 @@ int32_t ext_adi_motor_get(uint8_t smart_port, uint8_t adi_port);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an motor
  *
  * \param smart_port
@@ -381,7 +381,7 @@ typedef int32_t ext_adi_encoder_t;
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an encoder
  *
  * \param enc
@@ -397,7 +397,7 @@ int32_t ext_adi_encoder_get(ext_adi_encoder_t enc);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an encoder
  *
  * \param smart_port
@@ -423,7 +423,7 @@ ext_adi_encoder_t ext_adi_encoder_init(uint8_t smart_port, uint8_t adi_port_top,
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an encoder
  *
  * \param enc
@@ -439,7 +439,7 @@ int32_t ext_adi_encoder_reset(ext_adi_encoder_t enc);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an encoder
  *
  * \param enc
@@ -467,7 +467,7 @@ typedef int32_t ext_adi_ultrasonic_t;
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an ultrasonic
  *
  * \param ult
@@ -483,7 +483,7 @@ int32_t ext_adi_ultrasonic_get(ext_adi_ultrasonic_t ult);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an ultrasonic
  *
  * \param smart_port
@@ -505,7 +505,7 @@ ext_adi_ultrasonic_t ext_adi_ultrasonic_init(uint8_t smart_port, uint8_t adi_por
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as an ultrasonic
  *
  * \param ult
@@ -536,7 +536,7 @@ typedef int32_t ext_adi_gyro_t;
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a gyro
  *
  * \param gyro
@@ -556,7 +556,7 @@ double ext_adi_gyro_get(ext_adi_gyro_t gyro);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a gyro
  *
  * \param smart_port
@@ -577,7 +577,7 @@ ext_adi_gyro_t ext_adi_gyro_init(uint8_t smart_port, uint8_t adi_port, double mu
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a gyro
  *
  * \param gyro
@@ -593,7 +593,7 @@ int32_t ext_adi_gyro_reset(ext_adi_gyro_t gyro);
  *
  * This function uses the following values of errno when an error state is
  * reached:
- * ENXIO - The given value is not within the range of ADI Ports
+ * ENXIO - The given value is not within the range of ADI Ports, or the Smart Port is not within the valid range
  * EADDRINUSE - The port is not configured as a gyro
  *
  * \param gyro
@@ -603,33 +603,6 @@ int32_t ext_adi_gyro_reset(ext_adi_gyro_t gyro);
  * failed, setting errno.
  */
 int32_t ext_adi_gyro_shutdown(ext_adi_gyro_t gyro);
-
-#define SMART_PORT_BITS 16
-#define SMART_PORT_MASK ((1 << SMART_PORT_BITS) - 1)
-
-/**
- * Macro Description: Given a merged ports variable, it sets the smart port and adi port to the values inside the
- * int32_t.
- */
-#define get_ports(ports, smart_port, adi_port) \
-	{                                            \
-		uint32_t uport = (uint32_t)ports;          \
-		smart_port = uport & SMART_PORT_MASK;      \
-		adi_port = uport >> SMART_PORT_BITS;       \
-	}
-
-/**
- * Macro description: Given a port mask, it sets the smart port to the value masked inside the int_32.
- */
-#define get_smart_port(ports, smart_port) \
-	{                                       \
-		uint32_t uport = (uint32_t)ports;     \
-		smart_port = uport & SMART_PORT_MASK; \
-	}
-
-static inline uint32_t merge_adi_ports(uint8_t smart_port, uint8_t adi_port) {
-	return (adi_port << SMART_PORT_BITS) | smart_port;
-}
 
 #ifdef __cplusplus
 }  // namespace c
